+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+### ollama-c-lient-v0.0.2 - Under Dev./Testing
+
+DATE: 
+
+SEVERITY: low
+
+BUGS-FIXED:
+- N/A
+
+NEW-FEATURES:
+- N/A
+
+IMPROVEMENTS:
+- N/A
+
+OTHERS:
+- changes for alignment with [Ollama v0.9.0 changes](https://github.com/ollama/ollama/releases/tag/v0.9.0)
+
 ### ollama-c-lient-v0.0.1
 
 DATE: 2025/05/18